feat: add ownership check before schedule deletion and simplify role card button layout

+++ CORDIS/lib/providers/schedule/cloud_schedule_provider.dart
@@ -165,7 +165,10 @@ class CloudScheduleProvider extends ChangeNotifier {
     notifyListeners();
 
     try {
-      await _repo.deleteSchedule(scheduleId, userId);
+      if (userId == _schedules[scheduleId]!.ownerFirebaseId) {
+        await _repo.deleteSchedule(scheduleId, userId);
+      }
+
       _schedules.remove(scheduleId);
     } catch (e) {
       _error = e.toString();
